intensity: sign warnings fixes min(i,u)

+++ JHA/jackpotcoin.cu
@@ -93,7 +93,7 @@ extern "C" int scanhash_jackpot(int thr_id, uint32_t *pdata,
 		((uint32_t*)ptarget)[7] = 0x000f;
 
 	int throughput = opt_work_size ? opt_work_size : (1 << 20); // 256*4096
-	throughput = min(throughput, max_nonce - first_nonce);
+	throughput = min(throughput, (int)(max_nonce - first_nonce));
 
 	static bool init[8] = {0,0,0,0,0,0,0,0};
 	if (!init[thr_id])
