+++ qubit/qubit.cu
@@ -80,7 +80,7 @@ extern "C" int scanhash_qubit(int thr_id, uint32_t *pdata,
 	uint32_t endiandata[20];
 	const uint32_t first_nonce = pdata[19];
 	int throughput = opt_work_size ? opt_work_size : (1 << 19); // 256*256*8
-	throughput = min(throughput, max_nonce - first_nonce);
+	throughput = min(throughput, (int)(max_nonce - first_nonce));
 
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0x0000ff;
